+++ src/entry.cpp
@@ -39,19 +39,10 @@ AddonDefinition AddonDef{};
 std::filesystem::path AddonPath;
 ID3D11Device *pd3dDevice = nullptr;
 
-static bool ctrlQWasPressed = false;
 
-void ToggleShowWindowGW2_RotaHelper(const char *keybindIdentifier, bool isPressed)
+void ToggleShowWindowGW2_RotaHelper(const char *, bool)
 {
-    if (isPressed && !ctrlQWasPressed)
-    {
-        Settings::ToggleShowWindow(Globals::SettingsPath);
-        ctrlQWasPressed = true;
-    }
-    else if (!isPressed)
-    {
-        ctrlQWasPressed = false;
-    }
+    Settings::ToggleShowWindow(Globals::SettingsPath);
 }
 
 void RegisterQuickAccessShortcut()
@@ -64,7 +55,7 @@ void RegisterQuickAccessShortcut()
     Globals::APIDefs->InputBinds.RegisterWithString(
         KB_TOGGLE_GW2_RotaHelper,
         ToggleShowWindowGW2_RotaHelper,
-        "CTRL+Q");
+        "(null)");
 }
 
 void DeregisterQuickAccessShortcut()
@@ -170,12 +161,11 @@ void AddonLoad(AddonAPI *aApi)
 
     if (Globals::APIDefs && Globals::APIDefs->DataLink.Get)
     {
-        IDXGISwapChain *pSwapChain =
-            (IDXGISwapChain *)Globals::APIDefs->SwapChain;
+        auto *pSwapChain = (IDXGISwapChain *)Globals::APIDefs->SwapChain;
         if (pSwapChain)
         {
-            HRESULT hr = pSwapChain->GetDevice(__uuidof(ID3D11Device),
-                                               (void **)&pd3dDevice);
+            auto hr = pSwapChain->GetDevice(__uuidof(ID3D11Device),
+                                            (void **)&pd3dDevice);
             if (FAILED(hr))
                 pd3dDevice = nullptr;
         }
@@ -234,7 +224,7 @@ void AddonRender()
         profession = curr_profession;
     }
 
-    Globals::Render.toggle_vis(Settings::ShowWindow);
+    Globals::Render.set_show_window(Settings::ShowWindow);
     Globals::Render.render(pd3dDevice);
 }